@@ -75,7 +75,7 @@
     "classnames": "^2.2.5",
     "deepmerge": "^2.0.1",
     "dom-helpers": "^3.2.1",
-    "hoist-non-react-statics": "^2.3.1",
+    "hoist-non-react-statics": "^2.5.0",
     "jss": "^9.3.3",
     "jss-camel-case": "^6.0.0",
     "jss-default-unit": "^8.0.2",
@@ -89,6 +89,7 @@
     "prop-types": "^15.6.0",
     "react-event-listener": "^0.5.1",
     "react-jss": "^8.1.0",
+    "react-lifecycles-compat": "^1.0.2",
     "react-popper": "^0.8.0",
     "react-scrollbar-size": "^2.0.2",
     "react-transition-group": "^2.2.1",
@@ -173,10 +174,10 @@
     "prettier": "^1.8.2",
     "prismjs": "^1.8.4",
     "raw-loader": "^0.5.1",
-    "react": "^16.1.1",
+    "react": "^16.2.0",
     "react-autosuggest": "^9.3.2",
     "react-docgen": "^3.0.0-beta10",
-    "react-dom": "^16.1.1",
+    "react-dom": "^16.2.0",
     "react-inspector": "^2.2.2",
     "react-number-format": "^3.0.2",
     "react-redux": "^5.0.6",
@@ -192,7 +193,7 @@
     "redux-logger": "^3.0.6",
     "rimraf": "^2.6.2",
     "sinon": "^4.1.2",
-    "size-limit": "^0.15.1",
+    "size-limit": "0.14.1",
     "typescript": "^2.6.1",
     "url-loader": "^0.6.2",
     "vrtest": "^0.2.0",
