@@ -6,6 +6,7 @@ import { connect } from 'react-redux';
 import { MuiThemeProvider } from 'material-ui/styles';
 import Reboot from 'material-ui/Reboot';
 import JssProvider from 'react-jss/lib/JssProvider';
+import polyfill from 'react-lifecycles-compat';
 import getPageContext, { getTheme } from 'docs/src/modules/styles/getPageContext';
 import AppFrame from 'docs/src/modules/components/AppFrame';
 import { lightTheme, darkTheme, setPrismTheme } from 'docs/src/modules/utils/prism';
@@ -22,26 +23,47 @@ if (process.browser && !global.__INSERTION_POINT__) {
   }
 }
 
+function uiThemeSideEffect(uiTheme) {
+  setPrismTheme(uiTheme.paletteType === 'light' ? lightTheme : darkTheme);
+  document.body.dir = uiTheme.direction;
+}
+
 class AppWrapper extends React.Component {
-  componentWillMount() {
-    this.pageContext = this.props.pageContext || getPageContext();
+  static getDerivedStateFromProps(nextProps, prevState) {
+    if (typeof prevState.pageContext === 'undefined') {
+      return {
+        prevProps: nextProps,
+        pageContext: nextProps.pageContext || getPageContext(),
+      };
+    }
+
+    const { prevProps } = prevState;
+
+    if (
+      nextProps.uiTheme.paletteType !== prevProps.uiTheme.paletteType ||
+      nextProps.uiTheme.direction !== prevProps.uiTheme.direction
+    ) {
+      return {
+        prevProps: nextProps,
+        pageContext: {
+          ...prevState.pageContext,
+          theme: getTheme(nextProps.uiTheme),
+        },
+      };
+    }
+
+    return null;
   }
 
+  state = {};
+
   componentDidMount() {
+    uiThemeSideEffect(this.props.uiTheme);
+
     // Remove the server-side injected CSS.
     const jssStyles = document.querySelector('#jss-server-side');
     if (jssStyles && jssStyles.parentNode) {
       jssStyles.parentNode.removeChild(jssStyles);
-    }
-
-    if (this.props.uiTheme.paletteType === 'light') {
-      setPrismTheme(lightTheme);
-    } else {
-      setPrismTheme(darkTheme);
-    }
-
-    if (document.body) {
-      document.body.dir = this.props.uiTheme.direction;
     }
 
     if ('serviceWorker' in navigator && process.env.NODE_ENV === 'production') {
@@ -49,40 +71,21 @@ class AppWrapper extends React.Component {
     }
   }
 
-  componentWillReceiveProps(nextProps) {
-    if (
-      nextProps.uiTheme.paletteType !== this.props.uiTheme.paletteType ||
-      nextProps.uiTheme.direction !== this.props.uiTheme.direction
-    ) {
-      this.pageContext.theme = getTheme(nextProps.uiTheme);
-
-      if (nextProps.uiTheme.paletteType === 'light') {
-        setPrismTheme(lightTheme);
-      } else {
-        setPrismTheme(darkTheme);
-      }
-
-      if (document.body) {
-        document.body.dir = nextProps.uiTheme.direction;
-      }
-    }
+  componentDidUpdate() {
+    uiThemeSideEffect(this.props.uiTheme);
   }
-
-  context = null;
 
   render() {
     const { children } = this.props;
+    const { pageContext } = this.state;
 
     return (
       <JssProvider
-        jss={this.pageContext.jss}
-        registry={this.pageContext.sheetsRegistry}
-        generateClassName={this.pageContext.generateClassName}
+        jss={pageContext.jss}
+        registry={pageContext.sheetsRegistry}
+        generateClassName={pageContext.generateClassName}
       >
-        <MuiThemeProvider
-          theme={this.pageContext.theme}
-          sheetsManager={this.pageContext.sheetsManager}
-        >
+        <MuiThemeProvider theme={pageContext.theme} sheetsManager={pageContext.sheetsManager}>
           <Reboot />
           <AppFrame>{children}</AppFrame>
           <GoogleTag />
@@ -100,4 +103,4 @@ AppWrapper.propTypes = {
 
 export default connect(state => ({
   uiTheme: state.theme,
-}))(AppWrapper);
+}))(polyfill(AppWrapper));
