@@ -9,6 +9,9 @@ const styles = theme => ({
   root: {
     height: 180,
   },
+  container: {
+    display: 'flex',
+  },
   paper: {
     margin: theme.spacing.unit,
   },
@@ -39,13 +42,22 @@ class SimpleZoom extends React.Component {
     return (
       <div className={classes.root}>
         <Switch checked={checked} onChange={this.handleChange} aria-label="collapse" />
-        <Zoom in={checked} style={{ transitionDelay: checked ? 1000 : 0 }}>
-          <Paper elevation={4} className={classes.paper}>
-            <svg className={classes.svg}>
-              <polygon points="0,100 50,00, 100,100" className={classes.polygon} />
-            </svg>
-          </Paper>
-        </Zoom>
+        <div className={classes.container}>
+          <Zoom in={checked}>
+            <Paper elevation={4} className={classes.paper}>
+              <svg className={classes.svg}>
+                <polygon points="0,100 50,00, 100,100" className={classes.polygon} />
+              </svg>
+            </Paper>
+          </Zoom>
+          <Zoom in={checked} style={{ transitionDelay: checked ? 500 : 0 }}>
+            <Paper elevation={4} className={classes.paper}>
+              <svg className={classes.svg}>
+                <polygon points="0,100 50,00, 100,100" className={classes.polygon} />
+              </svg>
+            </Paper>
+          </Zoom>
+        </div>
       </div>
     );
   }
