@@ -7,6 +7,7 @@ import EventListener from 'react-event-listener';
 import debounce from 'lodash/debounce';
 import Transition from 'react-transition-group/Transition';
 import ownerWindow from 'dom-helpers/ownerWindow';
+import polyfill from 'react-lifecycles-compat';
 import withTheme from '../styles/withTheme';
 import { duration } from '../styles/transitions';
 import { reflow, getTransitionProps } from './utils';
@@ -69,9 +70,19 @@ export function setTranslateValue(props, node) {
  * It uses [react-transition-group](https://github.com/reactjs/react-transition-group) internally.
  */
 class Slide extends React.Component {
-  state = {
-    mounted: false,
-  };
+  static getDerivedStateFromProps(nextProps, prevState) {
+    if (typeof prevState.mounted === 'undefined') {
+      return {
+        mounted: false,
+      };
+    }
+
+    return {
+      mounted: true,
+    };
+  }
+
+  state = {};
 
   componentDidMount() {
     // state.mounted handle SSR, once the component is mounted, we need
@@ -81,12 +92,6 @@ class Slide extends React.Component {
       // otherwise component will be shown when in=false.
       this.updatePosition();
     }
-  }
-
-  componentWillReceiveProps() {
-    this.setState({
-      mounted: true,
-    });
   }
 
   componentDidUpdate(prevProps) {
@@ -299,4 +304,4 @@ Slide.defaultProps = {
   },
 };
 
-export default withTheme()(Slide);
+export default withTheme()(polyfill(Slide));
